requirements.txt
Adafruit-Blinka==6.4.2
adafruit-circuitpython-neopixel==6.0.3
adafruit-circuitpython-pypixelbuf==2.2.4
Adafruit-PlatformDetect==3.5.0
Adafruit-PureIO==1.1.8
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
sanic==21.3.2
